move amp-date-countdown
