+++ build-system/compile/bundles.config.extensions.json
@@ -299,7 +299,6 @@
     "version": "1.0",
     "latestVersion": "0.1",
     "options": {
-      "npm": true,
       "bento": true,
       "hasCss": true
     }
