+++ src/bento/components/amp-date-countdown/1.0/messages.js
