+++ src/bento/components/amp-date-countdown/1.0/test/test-component.js
